refactor: Global Widgets forder rename
