+++ dristi/lib/src/core/global_providers/network_status/network_status_provider.dart
@@ -1,0 +1,9 @@
+import 'package:dristi/src/core/global_providers/network_status/network_status_notifier.dart';
+import 'package:dristi/src/core/global_providers/network_status/network_status_state.dart';
+
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+final networkStatusProvider =
+    NotifierProvider<NetworkStatusNotifier, NetworkState>(
+  NetworkStatusNotifier.new,
+);
