+++ dristi/lib/src/core/global_widgets/error_screen.dart
