+++ dristi/lib/src/core/routes/app_router.dart
@@ -1,7 +1,7 @@
+import 'package:dristi/src/core/global_widgets/error_screen.dart';
 import 'package:dristi/src/core/routes/app_routes.dart';
 import 'package:dristi/src/core/routes/navigation_helper.dart';
 import 'package:dristi/src/core/utils/localization_ext.dart';
-import 'package:dristi/src/core/widgets/error_screen.dart';
 import 'package:dristi/src/features/destinations/presentation/pages/destinations_list_screen.dart';
 import 'package:dristi/src/features/districts/presentation/pages/districts_list_screen.dart';
 import 'package:dristi/src/features/home/home_screen/pages/home_screen.dart';
